+++ plugin/lib/service/assets.dart
@@ -130,7 +130,7 @@ class Assets with Service implements NotificationsListener {
 
   // Implementation
   @protected
-  String get assetsFileName => _assetsName;
+  String get cacheFileName => _assetsName;
 
   @protected
   Future<File?> getCacheFile() async {
@@ -138,7 +138,7 @@ class Assets with Service implements NotificationsListener {
     if ((assetsDir != null) && !await assetsDir.exists()) {
       await assetsDir.create(recursive: true);
     }
-    String? cacheFilePath = (assetsDir != null) ? join(assetsDir.path, assetsFileName) : null;
+    String? cacheFilePath = (assetsDir != null) ? join(assetsDir.path, cacheFileName) : null;
     return (cacheFilePath != null) ? File(cacheFilePath) : null;
   }
 
@@ -171,9 +171,12 @@ class Assets with Service implements NotificationsListener {
   }
 
   @protected
+  String get networkAssetName => _assetsName;
+
+  @protected
   Future<String?> loadContentStringFromNet() async {
     try {
-      http.Response? response = (Config().assetsUrl != null) ? await Network().get("${Config().assetsUrl}/$assetsFileName") : null;
+      http.Response? response = (Config().assetsUrl != null) ? await Network().get("${Config().assetsUrl}/$networkAssetName") : null;
       return ((response != null) && (response.statusCode == 200)) ? response.body : null;
     } catch (e) {
       debugPrint(e.toString());